Rename e2e test suites and add selfmonitor suite
