+++ test/e2e/logs/fluentbit/suite_test.go
@@ -1,0 +1,24 @@
+//go:build e2e
+
+package fluentbit
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/format"
+
+	"github.com/kyma-project/telemetry-manager/test/testkit/suite"
+)
+
+func TestSuite(t *testing.T) {
+	format.MaxDepth = suite.GomegaMaxDepth
+	format.MaxLength = suite.GomegaMaxLenght
+
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "E2E Suite - FluentBit Logs")
+}
+
+var _ = BeforeSuite(suite.BeforeSuiteFunc)
+var _ = AfterSuite(suite.AfterSuiteFunc)
