+++ test/e2e/metrics/suite_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/kyma-project/telemetry-manager/test/testkit/suite"
 )
 
-func TestMisc(t *testing.T) {
+func TestSuite(t *testing.T) {
 	format.MaxDepth = suite.GomegaMaxDepth
 	format.MaxLength = suite.GomegaMaxLenght
