+++ test/e2e/misc/suite_test.go
@@ -1,6 +1,6 @@
 //go:build e2e
 
-package otel
+package misc
 
 import (
 	"testing"
@@ -12,12 +12,12 @@ import (
 	"github.com/kyma-project/telemetry-manager/test/testkit/suite"
 )
 
-func TestMisc(t *testing.T) {
+func TestSuite(t *testing.T) {
 	format.MaxDepth = suite.GomegaMaxDepth
 	format.MaxLength = suite.GomegaMaxLenght
 
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "E2E Suite - Logs")
+	RunSpecs(t, "E2E Suite - Misc")
 }
 
 var _ = BeforeSuite(suite.BeforeSuiteFunc)
